In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int first = 0, second = 1, sum, i;
+	uint64_t first = 0, second = 1, sum, i;
 	
 	for (i = 1; i<= 98; i++)
 	{
